requirements.txt
# PDF extraction
pdfplumber>=0.11.0
pymupdf>=1.24.0

# OCR fallback (for scanned PDFs)
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.17.0

# LLM for agentic lookup and extraction fallback
openai>=1.0.0
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
